+++ runtime/common/src/precompile/mod.rs
@@ -49,9 +49,9 @@ where
 	fn execute(
 		address: H160,
 		input: &[u8],
-		target_gas: Option<usize>,
+		target_gas: Option<u64>,
 		context: &Context,
-	) -> Option<core::result::Result<(ExitSucceed, Vec<u8>, usize), ExitError>> {
+	) -> Option<core::result::Result<(ExitSucceed, Vec<u8>, u64), ExitError>> {
 		EthereumPrecompiles::execute(address, input, target_gas, context).or_else(|| {
 			if is_acala_precompile(address) && !PrecompileCallerFilter::is_allowed(context.caller) {
 				debug::debug!(target: "evm", "Precompile no permission");
@@ -80,9 +80,9 @@ mod tests {
 	impl Precompile for DummyPrecompile {
 		fn execute(
 			_input: &[u8],
-			_target_gas: Option<usize>,
+			_target_gas: Option<u64>,
 			_context: &Context,
-		) -> core::result::Result<(ExitSucceed, Vec<u8>, usize), ExitError> {
+		) -> core::result::Result<(ExitSucceed, Vec<u8>, u64), ExitError> {
 			Ok((ExitSucceed::Stopped, vec![], 0))
 		}
 	}
